pyproject.toml
[project]
name = "fastapi-jwt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.20.0",
    "fastapi>=0.115.0",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.9",
    "python-ulid>=2.7.0",
    "sqlalchemy>=2.0.35",
    "sqlmodel>=0.0.22",
    "uvicorn[standard]>=0.30.6",
]

[tool.uv]
dev-dependencies = [
    "isort>=5.13.2",
    "mypy>=1.11.2",
]
